+++ sipXportLib/include/os/OsNotification.h
@@ -1,10 +1,12 @@
 //
-// Copyright (C) 2004, 2005 Pingtel Corp.
-// 
+// Copyright (C) 2004-2006 SIPfoundry Inc.
+// Licensed by SIPfoundry under the LGPL license.
+//
+// Copyright (C) 2004-2006 Pingtel Corp.  All rights reserved.
+// Licensed to SIPfoundry under a Contributor Agreement.
 //
 // $$
-////////////////////////////////////////////////////////////////////////
-//////
+///////////////////////////////////////////////////////////////////////////////
 
 
 #ifndef _OsNotification_h_
@@ -13,7 +15,8 @@
 // SYSTEM INCLUDES
 
 // APPLICATION INCLUDES
-#include "os/OsStatus.h"
+#include <os/OsStatus.h>
+#include <os/OsIntTypes.h>
 
 // DEFINES
 // MACROS
@@ -42,7 +45,7 @@ public:
 
 /* ============================ MANIPULATORS ============================== */
 
-   virtual OsStatus signal(const int eventData) = 0;
+   virtual OsStatus signal(const intptr_t eventData) = 0;
      //:Signal the occurrence of the event
 
 /* ============================ ACCESSORS ================================= */
